+++ src/app/arb_validation/arb_validation.ts
@@ -4,6 +4,7 @@ export interface ArbValidationData {
   value: string;
   nParams: number;
   nParentheses: number;
+  nHtmlEntities: number;
 }
 
 export interface ArbValidation {
@@ -14,6 +15,7 @@ export enum InvalidType {
   keyNotFound = "Key does not exist",
   invalidParameters = "Incorrect number of parameters",
   invalidParentheses = "Incorrect number of parentheses",
+  undecodedHtmlEntityExists = "Undecoded html entity exists",
 }
 
 export interface ValidationResult {