+++ src/app/util/constant.ts
@@ -1,4 +1,4 @@
 export class Constant {
-  public static appName = "arb-translator"
-  public static homePage = "https://marketplace.visualstudio.com/items?itemName=DevStory.arb-translator"
+  public static appName = "arb-translator";
+  public static homePage = "https://marketplace.visualstudio.com/items?itemName=DevStory.arb-translator";
 }